Update on "[quant][refactor] Add test_workflow_module.py"
Summary:
we can put both fake quant module and observer module tests in the test_workflow_module.py

Test Plan:
python test/test_quantization.py

Reviewers:
.

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
